+++ ts/browser/hash-fn.ts
@@ -1,5 +1,9 @@
 import { BaseHashInput, IBaseHashOptions, inputToArray, defaultHashLength } from '../base/hash-fn';
-import { hash as rawHash } from '../../dist/wasm/browser/blake3_js';
+import {
+  hash as rawHash,
+  create_derive as createDerive,
+  create_keyed as createKeyed,
+} from '../../dist/wasm/browser/blake3_js';
 import { Hash } from './hash';
 
 /**
@@ -24,5 +28,41 @@ export function hash(
 ): Hash {
   const result = new Hash(length);
   rawHash(normalizeInput(input), result);
+  return result;
+}
+
+/**
+ * Given cryptographic key material  and a context string, services a subkey of
+ * any length. See {@link https://docs.rs/blake3/0.1.3/blake3/fn.derive_key.html}
+ * for more information.
+ */
+export function deriveKey(
+  context: string,
+  material: HashInput,
+  { length = defaultHashLength }: IBaseHashOptions = {},
+) {
+  const derive = createDerive(context);
+  derive.update(normalizeInput(material));
+  const result = new Hash(length);
+  derive.digest(result);
+  return result;
+}
+
+/**
+ * The keyed hash function. See {@link https://docs.rs/blake3/0.1.3/blake3/fn.keyed_hash.html}.
+ */
+export function keyedHash(
+  key: Buffer,
+  input: HashInput,
+  { length = defaultHashLength }: IBaseHashOptions = {},
+) {
+  if (key.length !== 32) {
+    throw new Error(`key provided to keyedHash must be 32 bytes, got ${key.length}`);
+  }
+
+  const derive = createKeyed(key);
+  derive.update(normalizeInput(input));
+  const result = new Hash(length);
+  derive.digest(result);
   return result;
 }