+++ ts/node/hash-instance.ts
@@ -109,3 +109,15 @@ export class NodeHash<Reader> extends Transform implements IHasher<Buffer> {
  * A Node.js crypto-like createHash method.
  */
 export const createHash = () => new NodeHash(getWasm().create_hasher(), r => new NodeHashReader(r));
+
+/**
+ * Construct a new Hasher for the keyed hash function.
+ */
+export const createKeyed = (key: Buffer) =>
+  new NodeHash(getWasm().create_keyed(key), r => new NodeHashReader(r));
+
+/**
+ * Construct a new Hasher for the key derivation function.
+ */
+export const createDeriveKey = (context: string) =>
+  new NodeHash(getWasm().create_derive(context), r => new NodeHashReader(r));