+++ ts/node/index.ts
@@ -1,4 +1,4 @@
-export { hash, HashInput } from './hash-fn';
+export { hash, deriveKey, keyedHash, HashInput } from './hash-fn';
 export * from './hash-instance';
 export * from './hash-reader';
 export * from '../base';